Respect and enforce PSR-12
