+++ .gitattributes
@@ -5,3 +5,4 @@
 /psalm.xml export-ignore
 /psalm-baseline.xml export-ignore
 /README.md export-ignore
+/phpcs.xml export-ignore
