+++ .github/workflows/quality.yml
@@ -28,3 +28,16 @@ jobs:
         run: composer install
       - name: Run tests
         run: vendor/bin/phpunit
+  code-style:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: shivammathur/setup-php@v2
+        with:
+          php-version: '8.1'
+          tools: composer:v2,codesniffer,cs2pr
+          coverage: none
+      - name: Install dependencies
+        run: composer install
+      - name: Check code style
+        run: phpcs --report=checkstyle | cs2pr
