+++ phpcs.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<ruleset>
+    <rule ref="PSR12">
+        <!-- we use constants to refer to functions, and functions are lowercase -->
+        <exclude name="Generic.NamingConventions.UpperCaseConstantName.ConstantNotUpperCase"/>
+        <!-- we support PHP 5.5, so we can't define constant visibility -->
+        <exclude name="PSR12.Properties.ConstantVisibility.NotFound"/>
+        <!-- we use Monad::of constants -->
+        <exclude name="Generic.NamingConventions.UpperCaseConstantName.ClassConstantNotUpperCase" />
+    </rule>
+</ruleset>
