+++ src/Exception/InvalidArgumentException.php
@@ -282,7 +282,11 @@ class InvalidArgumentException extends \InvalidArgumentException
     public static function assertPair($pair, $callee, $position)
     {
         if (!(is_array($pair) || $pair instanceof \ArrayAccess) || !isset($pair[0], $pair[1])) {
-            throw new static(\sprintf('%s() expects parameter %d to be a pair (array with two elements)', $callee, $position));
+            throw new static(\sprintf(
+                '%s() expects parameter %d to be a pair (array with two elements)',
+                $callee,
+                $position
+            ));
         }
     }
