+++ src/Identity.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Functional;
+
+class Identity extends Monad
+{
+    const of = "Functional\Identity::of";
+}
