+++ src/partial.php
@@ -16,7 +16,7 @@ namespace Functional;
 function partial(callable $f, $arg1)
 {
     $args = array_slice(func_get_args(), 1);
-    return function() use ($f, $args) {
+    return function () use ($f, $args) {
         return call_user_func_array($f, array_merge($args, func_get_args()));
     };
 }
@@ -36,7 +36,7 @@ define('Functional\partial', __NAMESPACE__ . '\\partial');
 function partial_r(callable $f, $arg1)
 {
     $args = array_slice(func_get_args(), 1);
-    return function() use ($f, $args) {
+    return function () use ($f, $args) {
         return call_user_func_array($f, array_merge(func_get_args(), $args));
     };
 }
@@ -53,7 +53,7 @@ define('Functional\partial_r', __NAMESPACE__ . '\\partial_r');
  */
 function partial_p(callable $f, array $args)
 {
-    return function() use ($f, $args) {
+    return function () use ($f, $args) {
         $_args = func_get_args();
         $position = 1;
         do {
@@ -61,7 +61,7 @@ function partial_p(callable $f, array $args)
                 $args[$position] = array_shift($_args);
             }
             ++$position;
-        } while($_args);
+        } while ($_args);
         ksort($args);
         return call_user_func_array($f, $args);
     };
